refactor login and auth error
